requirements.txt
notion==0.0.25
boto3
python-dotenv
